requirements.txt
alembic==1.14.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Flask==2.3.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.0.2
greenlet==3.1.1
gunicorn==21.2.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.8
MarkupSafe==3.0.2
packaging==24.2
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==2.3.7
